// src/app/generated/model/projectCreateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload to create a new project
 */
export interface ProjectCreateRequest { 
    /**
     * Unique short code for the project
     */
    code: string;
    /**
     * Project name
     */
    name: string;
    /**
     * Optional free-text description
     */
    description?: string;
    /**
     * Optional planned budget
     */
    budget?: number;
    /**
     * Date the project starts
     */
    startDate: string;
    /**
     * Optional planned/actual end date
     */
    endDate?: string;
}